tests/requirements.txt
# Test requirements for Cisco DME Collection
# Copyright 2025 Sagar Paul (@KB-perByte)
# GNU General Public License v3.0+

# Core testing dependencies
pytest>=6.0.0
pytest-mock>=3.6.0
pytest-cov>=2.12.0
pytest-xdist>=2.3.0

# Ansible dependencies
ansible-core>=2.16.0

# Network testing dependencies
requests>=2.25.0
urllib3>=1.26.0

# Additional test utilities
pyyaml>=5.4.0
jinja2>=3.0.0

# Development dependencies (optional)
black>=21.0.0
flake8>=3.9.0
pylint>=2.8.0
